GuidedTours: Rename dir to guided-tours, move to client/layout
Only directory name changing in this PR
