+++ assets/stylesheets/_components.scss
@@ -116,7 +116,7 @@
 @import 'components/vertical-nav/item/style';
 @import 'components/web-preview/style';
 @import 'devdocs/style';
-@import 'guidestours/style';
+@import 'layout/guided-tours/style';
 @import 'layout/community-translator/style';
 @import 'layout/masterbar/style';
 @import 'layout/nux-welcome/style';
