+++ client/layout/guided-tours/config.js
