+++ client/layout/guided-tours/index.js
