+++ client/layout/guided-tours/positioning.js
