+++ client/layout/guided-tours/steps.js
