+++ client/layout/guided-tours/style.scss
