+++ client/layout/index.jsx
@@ -22,7 +22,7 @@ var MasterbarLoggedIn = require( 'layout/masterbar/logged-in' ),
 	EmailVerificationNotice = require( 'components/email-verification/email-verification-notice' ),
 	Welcome = require( 'my-sites/welcome/welcome' ),
 	WelcomeMessage = require( 'layout/nux-welcome/welcome-message' ),
-	GuidesTours = require( 'guidestours' ),
+	GuidesTours = require( 'layout/guided-tours' ),
 	analytics = require( 'lib/analytics' ),
 	config = require( 'config' ),
 	connect = require( 'react-redux' ).connect,
